REF: single task + queue for udp packets to logstash
